+++ lib/screens/splash/splash.dart
@@ -1,6 +1,6 @@
 
-import 'package:cosmetic_app/splash/splash_animation1_element1.dart';
-import 'package:cosmetic_app/splash/splash_animation1_element2.dart';
+import 'package:cosmetic_app/screens/splash/splash_animation1_element1.dart';
+import 'package:cosmetic_app/screens/splash/splash_animation1_element2.dart';
 import 'package:cosmetic_app/values/values.dart';
 import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
