+++ lib/screens/splash/splash_animation1_element2.dart
