+++ lib/screens/tutorial1/tutorial1_animation1_element3.dart
