+++ lib/screens/tutorial2/tutorial2.dart
@@ -1,7 +1,7 @@
 
-import 'package:cosmetic_app/tutorial2/tutorial2_animation1_element1.dart';
-import 'package:cosmetic_app/tutorial2/tutorial2_animation1_element2.dart';
-import 'package:cosmetic_app/tutorial2/tutorial2_animation1_element3.dart';
+import 'package:cosmetic_app/screens/tutorial2/tutorial2_animation1_element1.dart';
+import 'package:cosmetic_app/screens/tutorial2/tutorial2_animation1_element2.dart';
+import 'package:cosmetic_app/screens/tutorial2/tutorial2_animation1_element3.dart';
 import 'package:cosmetic_app/values/values.dart';
 import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
