+++ lib/screens/tutorial3/tutorial3.dart
@@ -1,25 +1,23 @@
 
-import 'package:cosmetic_app/tutorial1/tutorial1_animation1_element1.dart';
-import 'package:cosmetic_app/tutorial1/tutorial1_animation1_element2.dart';
-import 'package:cosmetic_app/tutorial1/tutorial1_animation1_element3.dart';
-import 'package:cosmetic_app/tutorial1/tutorial1_animation1_element4.dart';
+import 'package:cosmetic_app/screens/tutorial3/tutorial3_animation1_element1.dart';
+import 'package:cosmetic_app/screens/tutorial3/tutorial3_animation1_element2.dart';
+import 'package:cosmetic_app/screens/tutorial3/tutorial3_animation1_element3.dart';
 import 'package:cosmetic_app/values/values.dart';
 import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
 
 
-class Tutorial1 extends StatefulWidget {
+class Tutorial3 extends StatefulWidget {
   
   @override
-  State<StatefulWidget> createState() => _Tutorial1State();
+  State<StatefulWidget> createState() => _Tutorial3State();
 }
 
 
-class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
+class _Tutorial3State extends State<Tutorial3> with TickerProviderStateMixin {
   AnimationController titleTwoAnimationController;
   AnimationController maintextTextAnimationController;
-  AnimationController faceimageImageAnimationController;
-  AnimationController colorimageImageAnimationController;
+  AnimationController selectlightfilterImageAnimationController;
   
   @override
   void initState() {
@@ -28,8 +26,7 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
     
     this.titleTwoAnimationController = AnimationController(duration: Duration(milliseconds: 500), vsync: this);
     this.maintextTextAnimationController = AnimationController(duration: Duration(milliseconds: 1000), vsync: this);
-    this.faceimageImageAnimationController = AnimationController(duration: Duration(milliseconds: 2500), vsync: this);
-    this.colorimageImageAnimationController = AnimationController(duration: Duration(milliseconds: 1500), vsync: this);
+    this.selectlightfilterImageAnimationController = AnimationController(duration: Duration(milliseconds: 1500), vsync: this);
     
     this.startAnimationOne();
   }
@@ -41,16 +38,14 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
     
     this.titleTwoAnimationController.dispose();
     this.maintextTextAnimationController.dispose();
-    this.faceimageImageAnimationController.dispose();
-    this.colorimageImageAnimationController.dispose();
+    this.selectlightfilterImageAnimationController.dispose();
   }
   
   void startAnimationOne() {
   
     this.titleTwoAnimationController.forward();
     this.maintextTextAnimationController.forward();
-    this.faceimageImageAnimationController.forward();
-    this.colorimageImageAnimationController.forward();
+    this.selectlightfilterImageAnimationController.forward();
   }
   
   @override
@@ -65,10 +60,10 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
         child: Column(
           children: [
             Container(
-              width: 116,
+              width: 147,
               height: 26,
               margin: EdgeInsets.only(top: 60),
-              child: Tutorial1WidgetAnimation1Element1(
+              child: Tutorial3WidgetAnimation1Element1(
                 animationController: this.titleTwoAnimationController,
                 child: Stack(
                   alignment: Alignment.center,
@@ -78,7 +73,7 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                       child: Opacity(
                         opacity: 0.5,
                         child: Container(
-                          width: 153,
+                          width: 180,
                           height: 30,
                           decoration: BoxDecoration(
                             color: Color.fromARGB(255, 254, 215, 2),
@@ -97,10 +92,10 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                           Align(
                             alignment: Alignment.centerLeft,
                             child: Text(
-                              "Step 1.",
+                              "Step 3.",
                               textAlign: TextAlign.left,
                               style: TextStyle(
-                                color: Color.fromARGB(255, 247, 7, 70),
+                                color: AppColors.primaryText,
                                 fontFamily: "Lobster 1.4",
                                 fontWeight: FontWeight.w400,
                                 fontSize: 20,
@@ -111,10 +106,10 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                           Align(
                             alignment: Alignment.centerLeft,
                             child: Text(
-                              "사진 찍기",
-                              textAlign: TextAlign.center,
+                              "필터 선택하기",
+                              textAlign: TextAlign.left,
                               style: TextStyle(
-                                color: Color.fromARGB(255, 247, 7, 70),
+                                color: AppColors.primaryText,
                                 fontFamily: "NanumBarunGothic",
                                 fontWeight: FontWeight.w700,
                                 fontSize: 16,
@@ -129,12 +124,12 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
               ),
             ),
             Container(
-              width: 230,
+              width: 243,
               margin: EdgeInsets.only(top: 24),
-              child: Tutorial1WidgetAnimation1Element2(
+              child: Tutorial3WidgetAnimation1Element2(
                 animationController: this.maintextTextAnimationController,
                 child: Text(
-                  "손목 발색 사진과\n민낯 사진을 업로드!",
+                  "장소에 따라\n바뀌는 색감을 확인",
                   textAlign: TextAlign.center,
                   style: TextStyle(
                     color: AppColors.accentText,
@@ -146,6 +141,7 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                 ),
               ),
             ),
+            Spacer(),
             Container(
               width: 280,
               height: 522,
@@ -159,7 +155,7 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                       width: 280,
                       height: 522,
                       decoration: BoxDecoration(
-                        color: Color.fromARGB(255, 54, 54, 54),
+                        color: AppColors.primaryBackground,
                         boxShadow: [
                           Shadows.primaryShadow,
                         ],
@@ -170,31 +166,17 @@ class _Tutorial1State extends State<Tutorial1> with TickerProviderStateMixin {
                   ),
                   Positioned(
                     top: 48,
-                    child: Container(
-                      width: 252,
-                      height: 448,
-                      decoration: BoxDecoration(
-                        color: Color.fromARGB(255, 0, 0, 0),
-                      ),
-                      child: Container(),
+                    child: Image.asset(
+                      "assets/images/screen_capture_3.png",
+                      fit: BoxFit.none,
                     ),
                   ),
                   Positioned(
                     top: 48,
-                    child: Tutorial1WidgetAnimation1Element4(
-                      animationController: this.colorimageImageAnimationController,
+                    child: Tutorial3WidgetAnimation1Element3(
+                      animationController: this.selectlightfilterImageAnimationController,
                       child: Image.asset(
-                        "assets/images/screen_capture_1.png",
-                        fit: BoxFit.none,
-                      ),
-                    ),
-                  ),
-                  Positioned(
-                    top: 48,
-                    child: Tutorial1WidgetAnimation1Element3(
-                      animationController: this.faceimageImageAnimationController,
-                      child: Image.asset(
-                        "assets/images/screen_capture_2.png",
+                        "assets/images/screen_capture_4.png",
                         fit: BoxFit.none,
                       ),
                     ),
